implement method getAllKnownServerNameAddress() and saveNewLed() in DB

+++ app/src/main/java/com/example/led_app/domain/Led.kt
@@ -7,13 +7,13 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "led")
 data class Led(
     @PrimaryKey(autoGenerate = true)
-    val uid: Int = 0,
+    val uid: Long = 0,
     @ColumnInfo(name = "led_name") val ledName: String?,
     @ColumnInfo(name = "ip_address") val ipAddress: String?,
 ) {
     companion object {
-        fun buildBaseOnLedData(ledName:String, ipAddress:String): Led {
-            return Led(ledName =  ledName, ipAddress = ipAddress)
+        fun buildBaseOnLedData(ledName: String, ipAddress: String): Led {
+            return Led(ledName = ledName, ipAddress = ipAddress)
         }
     }
